Add Step Decay Logic to terminal_gridsearch_relabelled.py
Moving the step decay support to the git version. This code has
been used extensively for evals, but wasn't committed and pushed!!
This code adds support for the step decay functionality. Includes
the arguments `--step_decay_epochs` to specify after how many epochs
to decay LR on, and `--step_decay_gamma` to specify the decay rate.

Signed-off-by: Sophie Haynes <s.c.haynes@rgu.ac.uk>