Undo Normalisation Modifications to terminal_gridsearch_relabelled.py
This commit removes the previous normalisation modifications. While the
current outputs are outwidth the expected ranges, they seem to behave
more stable and correctly. Various learning rates and momentum configurations
were evaluated with the [-1,1] normed images, however they consistently were
unable to decrease in loss nor increase in AUC internally (let alone external).

This may be due to the dataset distribution changing, however it is unclear at this time.
The code will be left in it's current state, given it seems to be learning reliably.

Signed-off-by: Sophie Haynes <s.c.haynes@rgu.ac.uk>